js/xpconnect/tests/mochitest/test_bug809547.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=809547
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 809547</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="go()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=809547">Mozilla Bug 809547</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 809547 **/
SimpleTest.waitForExplicitFinish();
var gObj = {};
function go() {
  window.location.expando = gObj;
  is(window.location.expando, gObj, "Expando appears");
  SimpleTest.executeSoon(finish);
}

function finish() {
  SpecialPowers.forceGC();
  SpecialPowers.forceCC();
  SpecialPowers.forceGC();
  SpecialPowers.forceCC();
  is(window.location.expando, gObj, "Expando preserved");
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
